requirements.txt
asgiref==3.5.2
Django==3.2.16
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0

django-filter~=22.1